Refactor: update kafka and events paths whitin common folder
