+++ src/app/Application.ts
@@ -2,7 +2,7 @@ import { CONSTANT_CONFIG } from '../core/constants/constantsConfig';
 import express, { Application as App } from 'express';
 import { MongoConnection } from '../core/db/MongoConnection';
 import { Logger } from '../utils/logger/Logger';
-import { RouterManager } from '../modules/common/router/RouterManager';
+import { RouterManager } from '../modules/common/express/router/RouterManager';
 import { inject, injectable } from 'inversify';
 import { TYPES } from '../core/inversify/types/inversifyTypes';
 import { userProducer } from '../modules/auth/producers/userProducer';
@@ -18,7 +18,9 @@ export class Application {
   private readonly PORT = CONSTANT_CONFIG.ENVIRONMENT.PORT;
   private readonly mongoConnection: MongoConnection;
 
-  constructor(@inject(TYPES.RouterManger) private readonly routerManager: RouterManager) {
+  constructor(
+    @inject(TYPES.RouterManger) private readonly routerManager: RouterManager
+  ) {
     this.app = express();
     this.mongoConnection = MongoConnection.getInstance();
   }
