+++ src/modules/common/express/response/ResponseManager.ts
@@ -1,10 +1,10 @@
 import { Response } from 'express';
-import { HttpStatusCode } from '../../../core/enums/HttpStatusCode';
-import { BaseException } from '../exceptions/BaseException';
-import { NotFoundException } from '../exceptions/NotFoundException';
-import { AlreadyExistException } from '../exceptions/AlreadyExistsException';
-import { UnauthorizedException } from '../exceptions/UnauthorizedException';
-import { CONSTANT_CONFIG } from '../../../core/constants/constantsConfig';
+import { HttpStatusCode } from '../../../../core/enums/HttpStatusCode';
+import { BaseException } from '../../exceptions/BaseException';
+import { NotFoundException } from '../../exceptions/NotFoundException';
+import { AlreadyExistException } from '../../exceptions/AlreadyExistsException';
+import { UnauthorizedException } from '../../exceptions/UnauthorizedException';
+import { CONSTANT_CONFIG } from '../../../../core/constants/constantsConfig';
 
 export class ResponseManager {
   constructor() {}
@@ -30,15 +30,21 @@ export class ResponseManager {
   ): Promise<Response> {
     if (error instanceof BaseException) {
       if (error instanceof NotFoundException) {
-        return appResponse.status(HttpStatusCode.NOT_FOUND).json({ message: error.message });
+        return appResponse
+          .status(HttpStatusCode.NOT_FOUND)
+          .json({ message: error.message });
       }
 
       if (error instanceof AlreadyExistException) {
-        return appResponse.status(HttpStatusCode.CONFLICT).json({ message: error.message });
+        return appResponse
+          .status(HttpStatusCode.CONFLICT)
+          .json({ message: error.message });
       }
 
       if (error instanceof UnauthorizedException) {
-        return appResponse.status(HttpStatusCode.UNAUTHORIZED).json({ message: error.message });
+        return appResponse
+          .status(HttpStatusCode.UNAUTHORIZED)
+          .json({ message: error.message });
       }
       return appResponse
         .status(HttpStatusCode.INTERNAL_SERVER_ERROR)
