+++ src/modules/common/express/router/RouterManager.ts
@@ -1,8 +1,8 @@
 import { Application } from 'express';
-import { UserRouter } from '../../auth/routes/UserRouter';
 import { inject, injectable } from 'inversify';
-import { TYPES } from '../../../core/inversify/types/inversifyTypes';
-import { ProductRouter } from '../../products/routes/ProductRouter';
+import { TYPES } from '../../../../core/inversify/types/inversifyTypes';
+import { UserRouter } from '../../../auth/routes/UserRouter';
+import { ProductRouter } from '../../../products/routes/ProductRouter';
 
 @injectable()
 export class RouterManager {
