+++ src/modules/common/kafka/events/interfaces/IEvent.ts
