+++ src/modules/notifications/consumer/notificationConsumer.ts
@@ -1,5 +1,5 @@
 import { kafka } from '../../../core/kafka/kafkaClient';
-import { CONSTANT_KAFKA } from '../../common/constants/constantsKafka';
+import { CONSTANT_KAFKA } from '../../common/kafka/constants/constantsKafka';
 import { mailerController } from '../mailer/controllers/mailerController';
 
 export const notificationConsumer = kafka.consumer({
